cli refactoring
